Warn if bullets contain notes
Bullets can contain notes with arbitrary text, which can be hard
to parse correctly.  This really should be handled more nicely, but
would require using a full grammar and parser, which is a much bigger
change.  That can be revisited in the future if needed; for now,
this doesn't change the program's limitations, it merely makes them
explicit.